iniciando correção do CollisionWorld

+++ src/math/rect.rs
@@ -1,13 +1,13 @@
 #[derive(Debug, Clone, Copy, PartialEq, PartialOrd)]
 pub struct Rect {
-    pub x: f32,
-    pub y: f32,
-    pub width: u32,
-    pub height: u32,
+    pub x: i32,
+    pub y: i32,
+    pub width: i32,
+    pub height: i32,
 }
 
 impl Rect {
-    pub fn new(x: f32, y: f32, width: u32, height: u32) -> Self {
+    pub fn new(x: i32, y: i32, width: i32, height: i32) -> Self {
         Self {
             x,
             y,
